fix: Resolve every conversion warning (#666)
Description:

Clears `-Wconversion`, `-Wsign-conversion` and `-Wfloat-conversion`:
**329 unique sites → 1**.

These are fixed at the type level rather than by casting. The recurring
finding was that a single concept was carried in several different types
along one path, and the warning fired wherever a value crossed between
them.

### Bug fixed

`BTOR2Encoder`'s `sorts_` and `terms_` were keyed by `int`, but
btor2parser reports line ids as `int64_t`. An input with ids past
`INT_MAX` had them truncated, so two distinct lines could silently alias
each other's term. `id2statenum` in the same file was already
`int64_t`-keyed. Committed separately as `65d75c7`.

### Types unified

- **Timesteps.** The unroller stored them as `size_t` but took them as
`unsigned int`. The engines carry bounds signed because `reached_k_`
uses `-1` for "nothing reached yet". `timestep()` and `prover_bound()`
in `prover.h` are now the two explicit crossings, and both throw rather
than wrap.
- **Widths.** `Sort::get_width()` returns `uint64_t`, but the BTOR2
encoder and the SyGuS syntax-analysis tables held widths in
`int`/`unsigned`, narrowing on the way in and widening on the way back
out.
- **Indices and counts.** Proof goals already stored their frame index
as `size_t` but were handed an `unsigned int`; cone-of-influence
verbosity now follows `Log`, which takes `size_t`.

### Behaviour changes

Values that used to wrap silently are now rejected:

- Widths and bit indices parsed from BTOR2 and SMV go through
`bit_index()` / `bv_index()`, which throw on a negative value instead of
widening it into an enormous index.
- `--random-seed` and `--ic3-gen-max-iter` are held as the `unsigned
int` smt-switch's reducer takes, and the parser refuses a value that
does not fit.
- The BMC and k-induction bound options, the value-abstraction cutoff
and the reset bound are range-checked at their boundaries.

### Loops

Three idioms turned out to be traps rather than style. `redxor` relied
on `width - 2` wrapping to a large unsigned value and converting back to
`-1` to terminate at width 1. Ten SMV printer loops and two others
walked vectors backwards through a signed index seeded from `size()`,
crossing the boundary twice per iteration; they use reverse iterators or
the count-down-from-size idiom now.

---------

+++ core/functional_unroller.cpp
@@ -38,7 +38,7 @@ FunctionalUnroller::FunctionalUnroller(const TransitionSystem & ts,
   }
 }
 
-Term FunctionalUnroller::at_time(const Term & t, unsigned int k)
+Term FunctionalUnroller::at_time(const Term & t, size_t k)
 {
   if (!ts_.no_next(t)) {
     throw PonoException(
@@ -47,14 +47,14 @@ Term FunctionalUnroller::at_time(const Term & t, unsigned int k)
   return super::at_time(t, k);
 }
 
-UnorderedTermMap & FunctionalUnroller::var_cache_at_time(unsigned int k)
+UnorderedTermMap & FunctionalUnroller::var_cache_at_time(size_t k)
 {
   const UnorderedTermMap & state_updates = ts_.state_updates();
   while (time_cache_.size() <= k) {
     time_cache_.push_back(UnorderedTermMap());
     extra_constraints_.push_back(true_);
     UnorderedTermMap & subst = time_cache_.back();
-    const unsigned int t = time_cache_.size() - 1;
+    const size_t t = time_cache_.size() - 1;
     assert(extra_constraints_.size() == t + 1);
 
     // create new state variables instead of substituting
